+++ pkgs/args/CHANGELOG.md
@@ -9,6 +9,8 @@
 * Remove sorting of the subcommands in usage output. Ordering will depend on the
   order that `addSubCommand` is called.
 * Remove extra newlines following separators when using flags without help text.
+* Throw an exception when parsing argument lists with missing mandatory
+  arguments instead of waiting until they are read.
 
 ## 2.7.0